backend/requirements.txt
Django>=3.0,<4.0
django-cors-headers
djangorestframework